backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.0
gunicorn>=21.0
